toolkit/content/tests/chrome/bug451286_window.xul

<?xml version="1.0"?>

<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is Findbar Test code
   -
   - The Initial Developer of the Original Code is
   - Graeme McCutcheon <graememcc_firefox@graeme-online.co.uk>.
   - Portions created by the Initial Developer are Copyright (C) 2008
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<window id="451286test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        width="600"
        height="600"
        onload="onLoad();"
        title="451286 test">

  <script type="application/javascript"><![CDATA[
    const Ci = Components.interfaces;
    const Cc = Components.classes;
    const Cr = Components.results;

    var gFindBar = null;
    var gBrowser;

    var imports = [ "SimpleTest", "ok", "snapshotWindow", "compareSnapshots"];
    for each (var import in imports) {
      window[import] = window.opener.wrappedJSObject[import];
    }

    function finish() {
      window.close();
      window.opener.wrappedJSObject.SimpleTest.finish();
    }

    function onLoad() {
      var _delayedOnLoad = function() {
        gFindBar = document.getElementById("FindToolbar");
        gBrowser = document.getElementById("content");
        gBrowser.addEventListener("pageshow", onPageShow, false);
        gBrowser.loadURI("data:text/html,<iframe id='if' src='data:text/html,text'></iframe>");
      }
      let tm = Cc["@mozilla.org/thread-manager;1"].
               getService(Ci.nsIThreadManager);
      tm.mainThread.dispatch({
        run: function() _delayedOnLoad()
      }, Ci.nsIThread.DISPATCH_NORMAL);
    }

    function onPageShow() {
      gBrowser.removeEventListener("pageshow", onPageShow, false);
      gFindBar.open();
      var search = "text";
      gFindBar._findField.value = search;
      var matchCase = gFindBar.getElement("find-case-sensitive");
      if (matchCase.checked)
        matchCase.doCommand();

      // Turn on highlighting
      gFindBar.toggleHighlight(true);
      gFindBar.close();
    
      // Take snapshot
      var gWin = gBrowser.contentWindow;
      var findSnapshot = snapshotWindow(gWin);

      // Now, perform the same selections manually
      gFindBar.open();
      gFindBar.toggleHighlight(false);
      gFindBar.close();

      var doc = gBrowser.contentDocument;
      var iframe = doc.getElementById("if");
      var ifBody = iframe.contentDocument.body;
      var range = iframe.contentDocument.createRange();
      range.selectNodeContents(ifBody.childNodes[0]);
      var ifWindow = gWin.frames[0];
      var ifDocShell = ifWindow.QueryInterface(Ci.nsIInterfaceRequestor)
                               .getInterface(Ci.nsIWebNavigation)
                               .QueryInterface(Ci.nsIDocShell);

      var ifController = ifDocShell.QueryInterface(Ci.nsIInterfaceRequestor)
                                   .getInterface(Ci.nsISelectionDisplay)
                                   .QueryInterface(Ci.nsISelectionController);

      var findSelection = ifController.getSelection(ifController.SELECTION_FIND);
      findSelection.addRange(range);

      //Take snapshots
      var manualSnapshot = snapshotWindow(gBrowser.contentWindow);
      var res = compareSnapshots(findSnapshot, manualSnapshot);
      ok(res[0], "Matches found in iframe correctly highlighted");
      finish();
    }
  ]]></script>

  <browser type="content-primary" flex="1" id="content" src="about:blank"/>
  <findbar id="FindToolbar" browserid="content"/>
</window>
